+++ python/plugins/processing/tests/CMakeLists.txt
@@ -16,7 +16,9 @@ IF(ENABLE_TESTS)
   ADD_PYTHON_TEST(ProcessingQgisAlgorithmsTestPt2 QgisAlgorithmsTest2.py)
   ADD_PYTHON_TEST(ProcessingQgisAlgorithmsTestPt3 QgisAlgorithmsTest3.py)
   ADD_PYTHON_TEST(ProcessingQgisAlgorithmsTestPt4 QgisAlgorithmsTest4.py)
-  ADD_PYTHON_TEST(ProcessingGdalAlgorithmsTest GdalAlgorithmsTest.py)
+  ADD_PYTHON_TEST(ProcessingGdalAlgorithmsGeneralTest GdalAlgorithmsGeneralTest.py)
+  ADD_PYTHON_TEST(ProcessingGdalAlgorithmsRasterTest GdalAlgorithmsRasterTest.py)
+  ADD_PYTHON_TEST(ProcessingGdalAlgorithmsVectorTest GdalAlgorithmsVectorTest.py)
   ADD_PYTHON_TEST(ProcessingGrass7AlgorithmsImageryTest Grass7AlgorithmsImageryTest.py)
   ADD_PYTHON_TEST(ProcessingGrass7AlgorithmsRasterTest Grass7AlgorithmsRasterTest.py)
   ADD_PYTHON_TEST(ProcessingGrass7AlgorithmsVectorTest Grass7AlgorithmsVectorTest.py)
