+++ python/plugins/processing/tests/GdalAlgorithmsRasterTest.py
@@ -2,7 +2,7 @@
 
 """
 ***************************************************************************
-    GdalAlgorithmTests.py
+    GdalAlgorithmRasterTest.py
     ---------------------
     Date                 : January 2016
     Copyright            : (C) 2016 by Matthias Kuhn
@@ -21,35 +21,36 @@ __author__ = 'Matthias Kuhn'
 __date__ = 'January 2016'
 __copyright__ = '(C) 2016, Matthias Kuhn'
 
-import AlgorithmsTestBase
-from processing.algs.gdal.OgrToPostGis import OgrToPostGis
-from processing.algs.gdal.GdalUtils import GdalUtils
+import nose2
+import os
+import shutil
+import tempfile
+
+from qgis.core import (QgsProcessingContext,
+                       QgsProcessingFeedback,
+                       QgsRectangle)
+
+from qgis.testing import (start_app,
+                          unittest)
+
 from processing.algs.gdal.AssignProjection import AssignProjection
-from processing.algs.gdal.Buffer import Buffer
 from processing.algs.gdal.ClipRasterByExtent import ClipRasterByExtent
 from processing.algs.gdal.ClipRasterByMask import ClipRasterByMask
-from processing.algs.gdal.Dissolve import Dissolve
-from processing.algs.gdal.gdal2tiles import gdal2tiles
-from processing.algs.gdal.gdalcalc import gdalcalc
-from processing.algs.gdal.gdaltindex import gdaltindex
-from processing.algs.gdal.contour import contour
-from processing.algs.gdal.gdalinfo import gdalinfo
+from processing.algs.gdal.ColorRelief import ColorRelief
 from processing.algs.gdal.GridAverage import GridAverage
 from processing.algs.gdal.GridDataMetrics import GridDataMetrics
 from processing.algs.gdal.GridInverseDistance import GridInverseDistance
 from processing.algs.gdal.GridInverseDistanceNearestNeighbor import GridInverseDistanceNearestNeighbor
 from processing.algs.gdal.GridLinear import GridLinear
 from processing.algs.gdal.GridNearestNeighbor import GridNearestNeighbor
-from processing.algs.gdal.buildvrt import buildvrt
+from processing.algs.gdal.gdal2tiles import gdal2tiles
+from processing.algs.gdal.gdalcalc import gdalcalc
+from processing.algs.gdal.gdaltindex import gdaltindex
+from processing.algs.gdal.contour import contour
+from processing.algs.gdal.gdalinfo import gdalinfo
 from processing.algs.gdal.hillshade import hillshade
 from processing.algs.gdal.aspect import aspect
-from processing.algs.gdal.ColorRelief import ColorRelief
-from processing.algs.gdal.ogr2ogr import ogr2ogr
-from processing.algs.gdal.ogrinfo import ogrinfo
-from processing.algs.gdal.OffsetCurve import OffsetCurve
-from processing.algs.gdal.OgrToPostGis import OgrToPostGis
-from processing.algs.gdal.OneSideBuffer import OneSideBuffer
-from processing.algs.gdal.PointsAlongLines import PointsAlongLines
+from processing.algs.gdal.buildvrt import buildvrt
 from processing.algs.gdal.proximity import proximity
 from processing.algs.gdal.rasterize import rasterize
 from processing.algs.gdal.retile import retile
@@ -66,34 +67,10 @@ from processing.algs.gdal.merge import merge
 from processing.algs.gdal.nearblack import nearblack
 from processing.algs.gdal.slope import slope
 
-from processing.tools.system import isWindows
-
-from qgis.core import (QgsProcessingContext,
-                       QgsProcessingFeedback,
-                       QgsCoordinateReferenceSystem,
-                       QgsApplication,
-                       QgsFeature,
-                       QgsGeometry,
-                       QgsPointXY,
-                       QgsProject,
-                       QgsVectorLayer,
-                       QgsRectangle,
-                       QgsProcessingException,
-                       QgsProcessingFeatureSourceDefinition)
-import nose2
-import os
-import shutil
-import tempfile
-
-from qgis.testing import (
-    start_app,
-    unittest
-)
-
 testDataPath = os.path.join(os.path.dirname(__file__), 'testdata')
 
 
-class TestGdalAlgorithms(unittest.TestCase, AlgorithmsTestBase.AlgorithmsTest):
+class TestGdalRasterAlgorithms(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
@@ -106,233 +83,6 @@ class TestGdalAlgorithms(unittest.TestCase, AlgorithmsTestBase.AlgorithmsTest):
     def tearDownClass(cls):
         for path in cls.cleanup_paths:
             shutil.rmtree(path)
-
-    def test_definition_file(self):
-        return 'gdal_algorithm_tests.yaml'
-
-    def testCommandName(self):
-        # Test that algorithms report a valid commandName
-        p = QgsApplication.processingRegistry().providerById('gdal')
-        for a in p.algorithms():
-            self.assertTrue(a.commandName(), 'Algorithm {} has no commandName!'.format(a.id()))
-
-    def testCommandNameInTags(self):
-        # Test that algorithms commandName is present in provided tags
-        p = QgsApplication.processingRegistry().providerById('gdal')
-        for a in p.algorithms():
-            self.assertTrue(a.commandName() in a.tags(), 'Algorithm {} commandName not found in tags!'.format(a.id()))
-
-    def testNoParameters(self):
-        # Test that algorithms throw QgsProcessingExceptions and not base Python
-        # exceptions when no parameters specified
-        p = QgsApplication.processingRegistry().providerById('gdal')
-        context = QgsProcessingContext()
-        feedback = QgsProcessingFeedback()
-        for a in p.algorithms():
-            try:
-                a.getConsoleCommands({}, context, feedback)
-            except QgsProcessingException:
-                pass
-
-    def testGetOgrCompatibleSourceFromMemoryLayer(self):
-        # create a memory layer and add to project and context
-        layer = QgsVectorLayer("Point?field=fldtxt:string&field=fldint:integer",
-                               "testmem", "memory")
-        self.assertTrue(layer.isValid())
-        pr = layer.dataProvider()
-        f = QgsFeature()
-        f.setAttributes(["test", 123])
-        f.setGeometry(QgsGeometry.fromPointXY(QgsPointXY(100, 200)))
-        f2 = QgsFeature()
-        f2.setAttributes(["test2", 457])
-        f2.setGeometry(QgsGeometry.fromPointXY(QgsPointXY(100, 200)))
-        self.assertTrue(pr.addFeatures([f, f2]))
-        self.assertEqual(layer.featureCount(), 2)
-        QgsProject.instance().addMapLayer(layer)
-        context = QgsProcessingContext()
-        context.setProject(QgsProject.instance())
-
-        alg = QgsApplication.processingRegistry().createAlgorithmById('gdal:buffervectors')
-        self.assertIsNotNone(alg)
-        parameters = {'INPUT': 'testmem'}
-        feedback = QgsProcessingFeedback()
-        # check that memory layer is automatically saved out to shape when required by GDAL algorithms
-        ogr_data_path, ogr_layer_name = alg.getOgrCompatibleSource('INPUT', parameters, context, feedback,
-                                                                   executing=True)
-        self.assertTrue(ogr_data_path)
-        self.assertTrue(ogr_data_path.endswith('.shp'))
-        self.assertTrue(os.path.exists(ogr_data_path))
-        self.assertTrue(ogr_layer_name)
-
-        # make sure that layer has correct features
-        res = QgsVectorLayer(ogr_data_path, 'res')
-        self.assertTrue(res.isValid())
-        self.assertEqual(res.featureCount(), 2)
-
-        # with memory layers - if not executing layer source should be ignored and replaced
-        # with a dummy path, because:
-        # - it has no meaning for the gdal command outside of QGIS, memory layers don't exist!
-        # - we don't want to force an export of the whole memory layer to a temp file just to show the command preview
-        # this might be very slow!
-        ogr_data_path, ogr_layer_name = alg.getOgrCompatibleSource('INPUT', parameters, context, feedback,
-                                                                   executing=False)
-        self.assertEqual(ogr_data_path, 'path_to_data_file')
-        self.assertEqual(ogr_layer_name, 'layer_name')
-
-        QgsProject.instance().removeMapLayer(layer)
-
-    def testGetOgrCompatibleSourceFromOgrLayer(self):
-        p = QgsProject()
-        source = os.path.join(testDataPath, 'points.gml')
-        vl = QgsVectorLayer(source)
-        self.assertTrue(vl.isValid())
-        p.addMapLayer(vl)
-
-        context = QgsProcessingContext()
-        context.setProject(p)
-        feedback = QgsProcessingFeedback()
-
-        alg = ogr2ogr()
-        alg.initAlgorithm()
-        path, layer = alg.getOgrCompatibleSource('INPUT', {'INPUT': vl.id()}, context, feedback, True)
-        self.assertEqual(path, source)
-        path, layer = alg.getOgrCompatibleSource('INPUT', {'INPUT': vl.id()}, context, feedback, False)
-        self.assertEqual(path, source)
-
-        # with selected features only - if not executing, the 'selected features only' setting
-        # should be ignored (because it has no meaning for the gdal command outside of QGIS!)
-        parameters = {'INPUT': QgsProcessingFeatureSourceDefinition(vl.id(), True)}
-        path, layer = alg.getOgrCompatibleSource('INPUT', parameters, context, feedback, False)
-        self.assertEqual(path, source)
-
-        # geopackage with layer
-        source = os.path.join(testDataPath, 'custom', 'circular_strings.gpkg')
-        vl2 = QgsVectorLayer(source + '|layername=circular_strings')
-        self.assertTrue(vl2.isValid())
-        p.addMapLayer(vl2)
-        path, layer = alg.getOgrCompatibleSource('INPUT', {'INPUT': vl2.id()}, context, feedback, True)
-        self.assertEqual(path, source)
-        self.assertEqual(layer, 'circular_strings')
-        vl3 = QgsVectorLayer(source + '|layername=circular_strings_with_line')
-        self.assertTrue(vl3.isValid())
-        p.addMapLayer(vl3)
-        path, layer = alg.getOgrCompatibleSource('INPUT', {'INPUT': vl3.id()}, context, feedback, True)
-        self.assertEqual(path, source)
-        self.assertEqual(layer, 'circular_strings_with_line')
-
-    def testGetOgrCompatibleSourceFromFeatureSource(self):
-        # create a memory layer and add to project and context
-        layer = QgsVectorLayer("Point?field=fldtxt:string&field=fldint:integer",
-                               "testmem", "memory")
-        self.assertTrue(layer.isValid())
-        pr = layer.dataProvider()
-        f = QgsFeature()
-        f.setAttributes(["test", 123])
-        f.setGeometry(QgsGeometry.fromPointXY(QgsPointXY(100, 200)))
-        f2 = QgsFeature()
-        f2.setAttributes(["test2", 457])
-        f2.setGeometry(QgsGeometry.fromPointXY(QgsPointXY(100, 200)))
-        self.assertTrue(pr.addFeatures([f, f2]))
-        self.assertEqual(layer.featureCount(), 2)
-        # select first feature
-        layer.selectByIds([next(layer.getFeatures()).id()])
-        self.assertEqual(len(layer.selectedFeatureIds()), 1)
-        QgsProject.instance().addMapLayer(layer)
-        context = QgsProcessingContext()
-        context.setProject(QgsProject.instance())
-
-        alg = QgsApplication.processingRegistry().createAlgorithmById('gdal:buffervectors')
-        self.assertIsNotNone(alg)
-        parameters = {'INPUT': QgsProcessingFeatureSourceDefinition('testmem', True)}
-        feedback = QgsProcessingFeedback()
-        # check that memory layer is automatically saved out to shape when required by GDAL algorithms
-        ogr_data_path, ogr_layer_name = alg.getOgrCompatibleSource('INPUT', parameters, context, feedback,
-                                                                   executing=True)
-        self.assertTrue(ogr_data_path)
-        self.assertTrue(ogr_data_path.endswith('.shp'))
-        self.assertTrue(os.path.exists(ogr_data_path))
-        self.assertTrue(ogr_layer_name)
-
-        # make sure that layer has only selected feature
-        res = QgsVectorLayer(ogr_data_path, 'res')
-        self.assertTrue(res.isValid())
-        self.assertEqual(res.featureCount(), 1)
-
-        QgsProject.instance().removeMapLayer(layer)
-
-    def testOgrOutputLayerName(self):
-        self.assertEqual(GdalUtils.ogrOutputLayerName('/home/me/out.shp'), 'out')
-        self.assertEqual(GdalUtils.ogrOutputLayerName('d:/test/test_out.shp'), 'test_out')
-        self.assertEqual(GdalUtils.ogrOutputLayerName('d:/test/TEST_OUT.shp'), 'TEST_OUT')
-        self.assertEqual(GdalUtils.ogrOutputLayerName('d:/test/test_out.gpkg'), 'test_out')
-
-    def testOgrLayerNameExtraction(self):
-        with tempfile.TemporaryDirectory() as outdir:
-            def _copyFile(dst):
-                shutil.copyfile(os.path.join(testDataPath, 'custom', 'grass7', 'weighted.csv'), dst)
-
-            # OGR provider - single layer
-            _copyFile(os.path.join(outdir, 'a.csv'))
-            name = GdalUtils.ogrLayerName(outdir)
-            self.assertEqual(name, 'a')
-
-            # OGR provider - multiple layers
-            _copyFile(os.path.join(outdir, 'b.csv'))
-            name1 = GdalUtils.ogrLayerName(outdir + '|layerid=0')
-            name2 = GdalUtils.ogrLayerName(outdir + '|layerid=1')
-            self.assertEqual(sorted([name1, name2]), ['a', 'b'])
-
-            name = GdalUtils.ogrLayerName(outdir + '|layerid=2')
-            self.assertIsNone(name)
-
-            # OGR provider - layername takes precedence
-            name = GdalUtils.ogrLayerName(outdir + '|layername=f')
-            self.assertEqual(name, 'f')
-
-            name = GdalUtils.ogrLayerName(outdir + '|layerid=0|layername=f')
-            self.assertEqual(name, 'f')
-
-            name = GdalUtils.ogrLayerName(outdir + '|layername=f|layerid=0')
-            self.assertEqual(name, 'f')
-
-            # SQLite provider
-            name = GdalUtils.ogrLayerName('dbname=\'/tmp/x.sqlite\' table="t" (geometry) sql=')
-            self.assertEqual(name, 't')
-
-            # PostgreSQL provider
-            name = GdalUtils.ogrLayerName(
-                'port=5493 sslmode=disable key=\'edge_id\' srid=0 type=LineString table="city_data"."edge" (geom) sql=')
-            self.assertEqual(name, 'city_data.edge')
-
-    def testOgrConnectionStringAndFormat(self):
-        context = QgsProcessingContext()
-        output, outputFormat = GdalUtils.ogrConnectionStringAndFormat('d:/test/test.shp', context)
-        self.assertEqual(output, 'd:/test/test.shp')
-        self.assertEqual(outputFormat, '"ESRI Shapefile"')
-        output, outputFormat = GdalUtils.ogrConnectionStringAndFormat('d:/test/test.mif', context)
-        self.assertEqual(output, 'd:/test/test.mif')
-        self.assertEqual(outputFormat, '"MapInfo File"')
-
-    def testCrsConversion(self):
-        self.assertFalse(GdalUtils.gdal_crs_string(QgsCoordinateReferenceSystem()))
-        self.assertEqual(GdalUtils.gdal_crs_string(QgsCoordinateReferenceSystem('EPSG:3111')), 'EPSG:3111')
-        self.assertEqual(GdalUtils.gdal_crs_string(QgsCoordinateReferenceSystem('POSTGIS:3111')), 'EPSG:3111')
-        self.assertEqual(GdalUtils.gdal_crs_string(QgsCoordinateReferenceSystem(
-            'proj4: +proj=utm +zone=36 +south +a=6378249.145 +b=6356514.966398753 +towgs84=-143,-90,-294,0,0,0,0 +units=m +no_defs')),
-            'EPSG:20936')
-        crs = QgsCoordinateReferenceSystem()
-        crs.createFromProj4(
-            '+proj=utm +zone=36 +south +a=600000 +b=70000 +towgs84=-143,-90,-294,0,0,0,0 +units=m +no_defs')
-        self.assertTrue(crs.isValid())
-        self.assertEqual(GdalUtils.gdal_crs_string(crs),
-                         '+proj=utm +zone=36 +south +a=600000 +b=70000 +towgs84=-143,-90,-294,0,0,0,0 +units=m +no_defs')
-        # check that newlines are stripped
-        crs = QgsCoordinateReferenceSystem()
-        crs.createFromProj4(
-            '+proj=utm +zone=36 +south\n     +a=600000 +b=70000 \r\n    +towgs84=-143,-90,-294,0,0,0,0 +units=m\n+no_defs')
-        self.assertTrue(crs.isValid())
-        self.assertEqual(GdalUtils.gdal_crs_string(crs),
-                         '+proj=utm +zone=36 +south      +a=600000 +b=70000       +towgs84=-143,-90,-294,0,0,0,0 +units=m +no_defs')
 
     def testAssignProjection(self):
         context = QgsProcessingContext()
@@ -374,47 +124,6 @@ class TestGdalAlgorithms(unittest.TestCase, AlgorithmsTestBase.AlgorithmsTest):
             ['gdal_edit.py',
              '-a_srs EPSG:3111 ' +
              source])
-
-    def testBuffer(self):
-        context = QgsProcessingContext()
-        feedback = QgsProcessingFeedback()
-        source = os.path.join(testDataPath, 'polys.gml')
-        source_with_space = os.path.join(testDataPath, 'filename with spaces.gml')
-        alg = Buffer()
-        alg.initAlgorithm()
-
-        with tempfile.TemporaryDirectory() as outdir:
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'DISTANCE': 5,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-dialect sqlite -sql "SELECT ST_Buffer(geometry, 5.0) AS geometry,* FROM \'polys2\'" ' +
-                 '-f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'DISTANCE': 5,
-                                        'DISSOLVE': True,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-dialect sqlite -sql "SELECT ST_Union(ST_Buffer(geometry, 5.0)) AS geometry,* FROM \'polys2\'" ' +
-                 '-f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'DISTANCE': 5,
-                                        'EXPLODE_COLLECTIONS': True,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-dialect sqlite -sql "SELECT ST_Buffer(geometry, 5.0) AS geometry,* FROM \'polys2\'" ' +
-                 '-explodecollections -f "ESRI Shapefile"'])
 
     def testGdalTranslate(self):
         context = QgsProcessingContext()
@@ -772,181 +481,6 @@ class TestGdalAlgorithms(unittest.TestCase, AlgorithmsTestBase.AlgorithmsTest):
                  source + ' ' +
                  outdir + '/check.shp'])
 
-    def testDissolve(self):
-        context = QgsProcessingContext()
-        feedback = QgsProcessingFeedback()
-        source = os.path.join(testDataPath, 'polys.gml')
-        source_with_space = os.path.join(testDataPath, 'filename with spaces.gml')
-        alg = Dissolve()
-        alg.initAlgorithm()
-
-        with tempfile.TemporaryDirectory() as outdir:
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-nlt PROMOTE_TO_MULTI -dialect sqlite -sql "SELECT ST_Union(geometry) AS geometry FROM \'polys2\'" ' +
-                 '-f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'FIELD': 'my_field',
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-nlt PROMOTE_TO_MULTI -dialect sqlite -sql "SELECT ST_Union(geometry) AS geometry, my_field FROM \'polys2\' ' +
-                 'GROUP BY my_field" -f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source_with_space,
-                                        'FIELD': 'my_field',
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 '"' + source_with_space + '" ' +
-                 '-nlt PROMOTE_TO_MULTI -dialect sqlite -sql "SELECT ST_Union(geometry) AS geometry, my_field FROM \'filename_with_spaces\' ' +
-                 'GROUP BY my_field" -f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'FIELD': 'my_field',
-                                        'GEOMETRY': 'the_geom',
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-nlt PROMOTE_TO_MULTI -dialect sqlite -sql "SELECT ST_Union(the_geom) AS the_geom, my_field FROM \'polys2\' ' +
-                 'GROUP BY my_field" -f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'FIELD': 'my_field',
-                                        'KEEP_ATTRIBUTES': False,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-nlt PROMOTE_TO_MULTI -dialect sqlite -sql "SELECT ST_Union(geometry) AS geometry, my_field FROM \'polys2\' ' +
-                 'GROUP BY my_field" -f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'KEEP_ATTRIBUTES': False,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-nlt PROMOTE_TO_MULTI -dialect sqlite -sql "SELECT ST_Union(geometry) AS geometry FROM \'polys2\'" ' +
-                 '-f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'FIELD': 'my_field',
-                                        'EXPLODE_COLLECTIONS': True,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-nlt PROMOTE_TO_MULTI -dialect sqlite -sql "SELECT ST_Union(geometry) AS geometry, my_field FROM \'polys2\' ' +
-                 'GROUP BY my_field" -explodecollections -f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'FIELD': 'my_field',
-                                        'COUNT_FEATURES': True,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-nlt PROMOTE_TO_MULTI -dialect sqlite -sql "SELECT ST_Union(geometry) AS geometry, my_field, COUNT(geometry) AS count FROM \'polys2\' ' +
-                 'GROUP BY my_field" -f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'FIELD': 'my_field',
-                                        'COUNT_FEATURES': True,
-                                        'GEOMETRY': 'the_geom',
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-nlt PROMOTE_TO_MULTI -dialect sqlite -sql "SELECT ST_Union(the_geom) AS the_geom, my_field, COUNT(the_geom) AS count FROM \'polys2\' ' +
-                 'GROUP BY my_field" -f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'FIELD': 'my_field',
-                                        'COMPUTE_AREA': True,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-nlt PROMOTE_TO_MULTI -dialect sqlite -sql "SELECT ST_Union(geometry) AS geometry, my_field, SUM(ST_Area(geometry)) AS area, ' +
-                 'ST_Perimeter(ST_Union(geometry)) AS perimeter FROM \'polys2\' ' +
-                 'GROUP BY my_field" -f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'FIELD': 'my_field',
-                                        'COMPUTE_AREA': True,
-                                        'GEOMETRY': 'the_geom',
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-nlt PROMOTE_TO_MULTI -dialect sqlite -sql "SELECT ST_Union(the_geom) AS the_geom, my_field, SUM(ST_Area(the_geom)) AS area, ' +
-                 'ST_Perimeter(ST_Union(the_geom)) AS perimeter FROM \'polys2\' ' +
-                 'GROUP BY my_field" -f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'FIELD': 'my_field',
-                                        'COMPUTE_STATISTICS': True,
-                                        'STATISTICS_ATTRIBUTE': 'my_val',
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-nlt PROMOTE_TO_MULTI -dialect sqlite -sql "SELECT ST_Union(geometry) AS geometry, my_field, ' +
-                 'SUM(my_val) AS sum, MIN(my_val) AS min, MAX(my_val) AS max, AVG(my_val) AS avg FROM \'polys2\' ' +
-                 'GROUP BY my_field" -f "ESRI Shapefile"'])
-
-            # compute stats without stats attribute, and vice versa (should be ignored)
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'FIELD': 'my_field',
-                                        'COMPUTE_STATISTICS': True,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-nlt PROMOTE_TO_MULTI -dialect sqlite -sql "SELECT ST_Union(geometry) AS geometry, my_field FROM \'polys2\' ' +
-                 'GROUP BY my_field" -f "ESRI Shapefile"'])
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'FIELD': 'my_field',
-                                        'STATISTICS_ATTRIBUTE': 'my_val',
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-nlt PROMOTE_TO_MULTI -dialect sqlite -sql "SELECT ST_Union(geometry) AS geometry, my_field FROM \'polys2\' ' +
-                 'GROUP BY my_field" -f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'FIELD': 'my_field',
-                                        'OPTIONS': 'my opts',
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-nlt PROMOTE_TO_MULTI -dialect sqlite -sql "SELECT ST_Union(geometry) AS geometry, my_field FROM \'polys2\' ' +
-                 'GROUP BY my_field" "my opts" -f "ESRI Shapefile"'])
-
     def testGdal2Tiles(self):
         context = QgsProcessingContext()
         feedback = QgsProcessingFeedback()
@@ -1041,7 +575,7 @@ class TestGdalAlgorithms(unittest.TestCase, AlgorithmsTestBase.AlgorithmsTest):
                                         'BAND_A': 1,
                                         'FORMULA': formula,
                                         'OUTPUT': output}, context, feedback),
-                ['gdal_calc' if isWindows() else 'gdal_calc.py',
+                ['gdal_calc.py',
                  '--calc "{}" --format JPEG --type Float32 -A {} --A_band 1 --outfile {}'.format(formula, source, output)])
 
             # check that formula is not escaped and formula is returned as it is
@@ -1051,7 +585,7 @@ class TestGdalAlgorithms(unittest.TestCase, AlgorithmsTestBase.AlgorithmsTest):
                                         'BAND_A': 1,
                                         'FORMULA': formula,
                                         'OUTPUT': output}, context, feedback),
-                ['gdal_calc' if isWindows() else 'gdal_calc.py',
+                ['gdal_calc.py',
                  '--calc "{}" --format JPEG --type Float32 -A {} --A_band 1 --outfile {}'.format(formula, source, output)])
 
             # additional creation options
@@ -1062,7 +596,7 @@ class TestGdalAlgorithms(unittest.TestCase, AlgorithmsTestBase.AlgorithmsTest):
                                         'FORMULA': formula,
                                         'OPTIONS': 'COMPRESS=JPEG|JPEG_QUALITY=75',
                                         'OUTPUT': output}, context, feedback),
-                ['gdal_calc' if isWindows() else 'gdal_calc.py',
+                ['gdal_calc.py',
                  '--calc "{}" --format JPEG --type Float32 -A {} --A_band 1 -co COMPRESS=JPEG -co JPEG_QUALITY=75 --outfile {}'.format(formula, source, output)])
 
             # additional parameters
@@ -1073,138 +607,8 @@ class TestGdalAlgorithms(unittest.TestCase, AlgorithmsTestBase.AlgorithmsTest):
                                         'FORMULA': formula,
                                         'EXTRA': '--debug --quiet',
                                         'OUTPUT': output}, context, feedback),
-                ['gdal_calc' if isWindows() else 'gdal_calc.py',
+                ['gdal_calc.py',
                  '--calc "{}" --format JPEG --type Float32 -A {} --A_band 1 --debug --quiet --outfile {}'.format(formula, source, output)])
-
-    def testBuildVrt(self):
-        context = QgsProcessingContext()
-        feedback = QgsProcessingFeedback()
-        source = os.path.join(testDataPath, 'dem.tif')
-        alg = buildvrt()
-        alg.initAlgorithm()
-
-        with tempfile.TemporaryDirectory() as outdir:
-            # defaults
-            cmd = alg.getConsoleCommands({'INPUT': [source],
-                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
-            t = cmd[1]
-            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
-            self.assertEqual(cmd,
-                             ['gdalbuildvrt',
-                              '-resolution average -separate -r nearest ' +
-                              '-input_file_list buildvrtInputFiles.txt ' +
-                              outdir + '/check.vrt'])
-            # custom resolution
-            cmd = alg.getConsoleCommands({'INPUT': [source],
-                                          'RESOLUTION': 2,
-                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
-            t = cmd[1]
-            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
-            self.assertEqual(cmd,
-                             ['gdalbuildvrt',
-                              '-resolution lowest -separate -r nearest ' +
-                              '-input_file_list buildvrtInputFiles.txt ' +
-                              outdir + '/check.vrt'])
-            # single layer
-            cmd = alg.getConsoleCommands({'INPUT': [source],
-                                          'SEPARATE': False,
-                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
-            t = cmd[1]
-            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
-            self.assertEqual(cmd,
-                             ['gdalbuildvrt',
-                              '-resolution average -r nearest ' +
-                              '-input_file_list buildvrtInputFiles.txt ' +
-                              outdir + '/check.vrt'])
-            # projection difference
-            cmd = alg.getConsoleCommands({'INPUT': [source],
-                                          'PROJ_DIFFERENCE': True,
-                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
-            t = cmd[1]
-            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
-            self.assertEqual(cmd,
-                             ['gdalbuildvrt',
-                              '-resolution average -separate -allow_projection_difference -r nearest ' +
-                              '-input_file_list buildvrtInputFiles.txt ' +
-                              outdir + '/check.vrt'])
-            # add alpha band
-            cmd = alg.getConsoleCommands({'INPUT': [source],
-                                          'ADD_ALPHA': True,
-                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
-            t = cmd[1]
-            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
-            self.assertEqual(cmd,
-                             ['gdalbuildvrt',
-                              '-resolution average -separate -addalpha -r nearest ' +
-                              '-input_file_list buildvrtInputFiles.txt ' +
-                              outdir + '/check.vrt'])
-            # assign CRS
-            cmd = alg.getConsoleCommands({'INPUT': [source],
-                                          'ASSIGN_CRS': 'EPSG:3111',
-                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
-            t = cmd[1]
-            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
-            self.assertEqual(cmd,
-                             ['gdalbuildvrt',
-                              '-resolution average -separate -a_srs EPSG:3111 -r nearest ' +
-                              '-input_file_list buildvrtInputFiles.txt ' +
-                              outdir + '/check.vrt'])
-
-            custom_crs = 'proj4: +proj=utm +zone=36 +south +a=6378249.145 +b=6356514.966398753 +towgs84=-143,-90,-294,0,0,0,0 +units=m +no_defs'
-            cmd = alg.getConsoleCommands({'INPUT': [source],
-                                          'ASSIGN_CRS': custom_crs,
-                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
-            t = cmd[1]
-            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
-            self.assertEqual(cmd,
-                             ['gdalbuildvrt',
-                              '-resolution average -separate -a_srs EPSG:20936 -r nearest ' +
-                              '-input_file_list buildvrtInputFiles.txt ' +
-                              outdir + '/check.vrt'])
-            # source NODATA
-            cmd = alg.getConsoleCommands({'INPUT': [source],
-                                          'SRC_NODATA': '-9999',
-                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
-            t = cmd[1]
-            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
-            self.assertEqual(cmd,
-                             ['gdalbuildvrt',
-                              '-resolution average -separate -r nearest -srcnodata "-9999" ' +
-                              '-input_file_list buildvrtInputFiles.txt ' +
-                              outdir + '/check.vrt'])
-
-            cmd = alg.getConsoleCommands({'INPUT': [source],
-                                          'SRC_NODATA': '-9999 9999',
-                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
-            t = cmd[1]
-            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
-            self.assertEqual(cmd,
-                             ['gdalbuildvrt',
-                              '-resolution average -separate -r nearest -srcnodata "-9999 9999" ' +
-                              '-input_file_list buildvrtInputFiles.txt ' +
-                              outdir + '/check.vrt'])
-
-            cmd = alg.getConsoleCommands({'INPUT': [source],
-                                          'SRC_NODATA': '',
-                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
-            t = cmd[1]
-            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
-            self.assertEqual(cmd,
-                             ['gdalbuildvrt',
-                              '-resolution average -separate -r nearest ' +
-                              '-input_file_list buildvrtInputFiles.txt ' +
-                              outdir + '/check.vrt'])
-            # additional parameters
-            cmd = alg.getConsoleCommands({'INPUT': [source],
-                                          'EXTRA': '-overwrite -optim RASTER -vrtnodata -9999',
-                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
-            t = cmd[1]
-            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
-            self.assertEqual(cmd,
-                             ['gdalbuildvrt',
-                              '-resolution average -separate -r nearest -overwrite -optim RASTER -vrtnodata -9999 ' +
-                              '-input_file_list buildvrtInputFiles.txt ' +
-                              outdir + '/check.vrt'])
 
     def testGdalInfo(self):
         context = QgsProcessingContext()
@@ -1611,423 +1015,6 @@ class TestGdalAlgorithms(unittest.TestCase, AlgorithmsTestBase.AlgorithmsTest):
                  '-z_multiply 1.5 -outsize 1754 1394 ' +
                  source + ' ' +
                  outdir + '/check.tif'])
-
-    def testOgr2Ogr(self):
-        context = QgsProcessingContext()
-        feedback = QgsProcessingFeedback()
-        source = os.path.join(testDataPath, 'polys.gml')
-        alg = ogr2ogr()
-        alg.initAlgorithm()
-
-        with tempfile.TemporaryDirectory() as outdir:
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 '-f "ESRI Shapefile" ' + outdir + '/check.shp ' +
-                 source + ' polys2'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'OUTPUT': outdir + '/check.kml'}, context, feedback),
-                ['ogr2ogr',
-                 '-f "LIBKML" ' + outdir + '/check.kml ' +
-                 source + ' polys2'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'OUTPUT': outdir + '/my out/check.kml'}, context, feedback),
-                ['ogr2ogr',
-                 '-f "LIBKML" "' + outdir + '/my out/check.kml" ' +
-                 source + ' polys2'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'OUTPUT': outdir + '/check.gpkg'}, context, feedback),
-                ['ogr2ogr',
-                 '-f "GPKG" ' + outdir + '/check.gpkg ' +
-                 source + ' polys2'])
-
-    def testOgr2PostGis(self):
-        context = QgsProcessingContext()
-        feedback = QgsProcessingFeedback()
-        source = os.path.join(testDataPath, 'polys.gml')
-        source_with_space = os.path.join(testDataPath, 'filename with spaces.gml')
-        alg = OgrToPostGis()
-        alg.initAlgorithm()
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source_with_space}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 "' + source_with_space + '" filename_with_spaces '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.filename_with_spaces -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'HOST': 'google.com'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=google.com port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'PORT': 3333}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=3333 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'USER': 'kevin_bacon'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public user=kevin_bacon" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'DBNAME': 'secret_stuff'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 dbname=secret_stuff active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'PASSWORD': 'passw0rd'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 password=passw0rd active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'SCHEMA': 'desktop'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=desktop" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln desktop.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'TABLE': 'out_table'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.out_table -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'PK': ''}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'PK': 'new_fid'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=new_fid -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'PK': '',
-                                    'PRIMARY_KEY': 'objectid'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=objectid -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'PK': 'new_id',
-                                    'PRIMARY_KEY': 'objectid'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=new_id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'GEOCOLUMN': 'my_geom'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=my_geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'DIM': 1}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=3 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'SIMPLIFY': 5}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -simplify 5 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'SEGMENTIZE': 4}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -segmentize 4 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'SPAT': QgsRectangle(1, 2, 3, 4)}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -spat 1.0 2.0 3.0 4.0 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'FIELDS': ['f1', 'f2']}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 -select "f1,f2" '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'WHERE': '0=1'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -where "0=1" -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'GT': 2}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -gt 2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'OVERWRITE': False}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'APPEND': True}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-append -overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'ADDFIELDS': True}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-addfields -overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'LAUNDER': True}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-lco LAUNDER=NO -overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'INDEX': True}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-lco SPATIAL_INDEX=OFF -overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'SKIPFAILURES': True}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -skipfailures -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'PROMOTETOMULTI': False}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'PRECISION': False}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI -lco PRECISION=NO'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'OPTIONS': 'blah'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI blah'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'SHAPE_ENCODING': 'blah'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES --config SHAPE_ENCODING "blah" -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'GTYPE': 4}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -nlt LINESTRING -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'A_SRS': 'EPSG:3111'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -a_srs EPSG:3111 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'A_SRS': QgsCoordinateReferenceSystem('EPSG:3111')}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -a_srs EPSG:3111 -nlt PROMOTE_TO_MULTI'])
-
-        custom_crs = 'proj4: +proj=utm +zone=36 +south +a=6378249.145 +b=6356514.966398753 +towgs84=-143,-90,-294,0,0,0,0 +units=m +no_defs'
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'A_SRS': custom_crs}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -a_srs EPSG:20936 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'T_SRS': 'EPSG:3111'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -t_srs EPSG:3111 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'T_SRS': QgsCoordinateReferenceSystem('EPSG:3111')}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -t_srs EPSG:3111 -nlt PROMOTE_TO_MULTI'])
-
-        custom_crs = 'proj4: +proj=utm +zone=36 +south +a=6378249.145 +b=6356514.966398753 +towgs84=-143,-90,-294,0,0,0,0 +units=m +no_defs'
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'T_SRS': custom_crs}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -t_srs EPSG:20936 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'S_SRS': 'EPSG:3111'}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -s_srs EPSG:3111 -nlt PROMOTE_TO_MULTI'])
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'S_SRS': QgsCoordinateReferenceSystem('EPSG:3111')}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -s_srs EPSG:3111 -nlt PROMOTE_TO_MULTI'])
-
-        custom_crs = 'proj4: +proj=utm +zone=36 +south +a=6378249.145 +b=6356514.966398753 +towgs84=-143,-90,-294,0,0,0,0 +units=m +no_defs'
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'S_SRS': custom_crs}, context, feedback),
-            ['ogr2ogr',
-             '-progress --config PG_USE_COPY YES -f PostgreSQL "PG:host=localhost port=5432 active_schema=public" '
-             '-lco DIM=2 ' + source + ' polys2 '
-             '-overwrite -lco GEOMETRY_NAME=geom -lco FID=id -nln public.polys2 -s_srs EPSG:20936 -nlt PROMOTE_TO_MULTI'])
-
-    def testOgrInfo(self):
-        context = QgsProcessingContext()
-        feedback = QgsProcessingFeedback()
-        source = os.path.join(testDataPath, 'polys.gml')
-        alg = ogrinfo()
-        alg.initAlgorithm()
-
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'SUMMARY_ONLY': True,
-                                    'NO_METADATA': False}, context, feedback),
-            ['ogrinfo',
-             '-al -so ' +
-             source])
-
-        source = os.path.join(testDataPath, 'filename with spaces.gml')
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'SUMMARY_ONLY': True,
-                                    'NO_METADATA': False}, context, feedback),
-            ['ogrinfo',
-             '-al -so "' +
-             source + '"'])
-
-        source = os.path.join(testDataPath, 'filename with spaces.gml')
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'SUMMARY_ONLY': False,
-                                    'NO_METADATA': False}, context, feedback),
-            ['ogrinfo',
-             '-al "' +
-             source + '"'])
-
-        source = os.path.join(testDataPath, 'filename with spaces.gml')
-        self.assertEqual(
-            alg.getConsoleCommands({'INPUT': source,
-                                    'SUMMARY_ONLY': True,
-                                    'NO_METADATA': True}, context, feedback),
-            ['ogrinfo',
-             '-al -so -nomd "' +
-             source + '"'])
 
     def testHillshade(self):
         context = QgsProcessingContext()
@@ -2962,85 +1949,6 @@ class TestGdalAlgorithms(unittest.TestCase, AlgorithmsTestBase.AlgorithmsTest):
                  source + ' ' +
                  outsource])
 
-    def testOffsetCurve(self):
-        context = QgsProcessingContext()
-        feedback = QgsProcessingFeedback()
-        source = os.path.join(testDataPath, 'polys.gml')
-        source_with_space = os.path.join(testDataPath, 'filename with spaces.gml')
-        alg = OffsetCurve()
-        alg.initAlgorithm()
-
-        with tempfile.TemporaryDirectory() as outdir:
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'DISTANCE': 5,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-dialect sqlite -sql "SELECT ST_OffsetCurve(geometry, 5.0) AS geometry,* FROM \'polys2\'" ' +
-                 '-f "ESRI Shapefile"'])
-
-    def testOneSidedBuffer(self):
-        context = QgsProcessingContext()
-        feedback = QgsProcessingFeedback()
-        source = os.path.join(testDataPath, 'polys.gml')
-        source_with_space = os.path.join(testDataPath, 'filename with spaces.gml')
-        alg = OneSideBuffer()
-        alg.initAlgorithm()
-
-        with tempfile.TemporaryDirectory() as outdir:
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'DISTANCE': 5,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-dialect sqlite -sql "SELECT ST_SingleSidedBuffer(geometry, 5.0, 0) AS geometry,* FROM \'polys2\'" ' +
-                 '-f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'DISTANCE': 5,
-                                        'DISSOLVE': True,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-dialect sqlite -sql "SELECT ST_Union(ST_SingleSidedBuffer(geometry, 5.0, 0)) AS geometry,* FROM \'polys2\'" ' +
-                 '-f "ESRI Shapefile"'])
-
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'DISTANCE': 5,
-                                        'EXPLODE_COLLECTIONS': True,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-dialect sqlite -sql "SELECT ST_SingleSidedBuffer(geometry, 5.0, 0) AS geometry,* FROM \'polys2\'" ' +
-                 '-explodecollections -f "ESRI Shapefile"'])
-
-    def testPointsAlongLines(self):
-        context = QgsProcessingContext()
-        feedback = QgsProcessingFeedback()
-        source = os.path.join(testDataPath, 'polys.gml')
-        source_with_space = os.path.join(testDataPath, 'filename with spaces.gml')
-        alg = PointsAlongLines()
-        alg.initAlgorithm()
-
-        with tempfile.TemporaryDirectory() as outdir:
-            self.assertEqual(
-                alg.getConsoleCommands({'INPUT': source,
-                                        'DISTANCE': 0.2,
-                                        'OUTPUT': outdir + '/check.shp'}, context, feedback),
-                ['ogr2ogr',
-                 outdir + '/check.shp ' +
-                 source + ' ' +
-                 '-dialect sqlite -sql "SELECT ST_Line_Interpolate_Point(geometry, 0.2) AS geometry,* FROM \'polys2\'" ' +
-                 '-f "ESRI Shapefile"'])
-
     def testGdalAddo(self):
         context = QgsProcessingContext()
         feedback = QgsProcessingFeedback()
@@ -3346,56 +2254,135 @@ class TestGdalAlgorithms(unittest.TestCase, AlgorithmsTestBase.AlgorithmsTest):
                  '-r cubic -of GTiff -bitdepth 12 -threads ALL_CPUS'
                  ])
 
-
-class TestGdalOgrToPostGis(unittest.TestCase):
-
-    @classmethod
-    def setUpClass(cls):
-        # start_app()
-        from processing.core.Processing import Processing
-        Processing.initialize()
-
-    @classmethod
-    def tearDownClass(cls):
-        pass
-
-    # See https://github.com/qgis/QGIS/issues/23629
-    def test_getConnectionString(self):
-        obj = OgrToPostGis()
-        obj.initAlgorithm({})
-
-        parameters = {}
+    def testBuildVrt(self):
         context = QgsProcessingContext()
+        feedback = QgsProcessingFeedback()
+        source = os.path.join(testDataPath, 'dem.tif')
+        alg = buildvrt()
+        alg.initAlgorithm()
 
-        # NOTE: defaults are debatable, see
-        # https://github.com/qgis/QGIS/pull/3607#issuecomment-253971020
-        self.assertEqual(obj.getConnectionString(parameters, context),
-                         "host=localhost port=5432 active_schema=public")
+        with tempfile.TemporaryDirectory() as outdir:
+            # defaults
+            cmd = alg.getConsoleCommands({'INPUT': [source],
+                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
+            t = cmd[1]
+            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
+            self.assertEqual(cmd,
+                             ['gdalbuildvrt',
+                              '-resolution average -separate -r nearest ' +
+                              '-input_file_list buildvrtInputFiles.txt ' +
+                              outdir + '/check.vrt'])
+            # custom resolution
+            cmd = alg.getConsoleCommands({'INPUT': [source],
+                                          'RESOLUTION': 2,
+                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
+            t = cmd[1]
+            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
+            self.assertEqual(cmd,
+                             ['gdalbuildvrt',
+                              '-resolution lowest -separate -r nearest ' +
+                              '-input_file_list buildvrtInputFiles.txt ' +
+                              outdir + '/check.vrt'])
+            # single layer
+            cmd = alg.getConsoleCommands({'INPUT': [source],
+                                          'SEPARATE': False,
+                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
+            t = cmd[1]
+            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
+            self.assertEqual(cmd,
+                             ['gdalbuildvrt',
+                              '-resolution average -r nearest ' +
+                              '-input_file_list buildvrtInputFiles.txt ' +
+                              outdir + '/check.vrt'])
+            # projection difference
+            cmd = alg.getConsoleCommands({'INPUT': [source],
+                                          'PROJ_DIFFERENCE': True,
+                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
+            t = cmd[1]
+            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
+            self.assertEqual(cmd,
+                             ['gdalbuildvrt',
+                              '-resolution average -separate -allow_projection_difference -r nearest ' +
+                              '-input_file_list buildvrtInputFiles.txt ' +
+                              outdir + '/check.vrt'])
+            # add alpha band
+            cmd = alg.getConsoleCommands({'INPUT': [source],
+                                          'ADD_ALPHA': True,
+                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
+            t = cmd[1]
+            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
+            self.assertEqual(cmd,
+                             ['gdalbuildvrt',
+                              '-resolution average -separate -addalpha -r nearest ' +
+                              '-input_file_list buildvrtInputFiles.txt ' +
+                              outdir + '/check.vrt'])
+            # assign CRS
+            cmd = alg.getConsoleCommands({'INPUT': [source],
+                                          'ASSIGN_CRS': 'EPSG:3111',
+                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
+            t = cmd[1]
+            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
+            self.assertEqual(cmd,
+                             ['gdalbuildvrt',
+                              '-resolution average -separate -a_srs EPSG:3111 -r nearest ' +
+                              '-input_file_list buildvrtInputFiles.txt ' +
+                              outdir + '/check.vrt'])
 
-        parameters['HOST'] = 'remote'
-        self.assertEqual(obj.getConnectionString(parameters, context),
-                         "host=remote port=5432 active_schema=public")
+            custom_crs = 'proj4: +proj=utm +zone=36 +south +a=6378249.145 +b=6356514.966398753 +towgs84=-143,-90,-294,0,0,0,0 +units=m +no_defs'
+            cmd = alg.getConsoleCommands({'INPUT': [source],
+                                          'ASSIGN_CRS': custom_crs,
+                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
+            t = cmd[1]
+            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
+            self.assertEqual(cmd,
+                             ['gdalbuildvrt',
+                              '-resolution average -separate -a_srs EPSG:20936 -r nearest ' +
+                              '-input_file_list buildvrtInputFiles.txt ' +
+                              outdir + '/check.vrt'])
+            # source NODATA
+            cmd = alg.getConsoleCommands({'INPUT': [source],
+                                          'SRC_NODATA': '-9999',
+                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
+            t = cmd[1]
+            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
+            self.assertEqual(cmd,
+                             ['gdalbuildvrt',
+                              '-resolution average -separate -r nearest -srcnodata "-9999" ' +
+                              '-input_file_list buildvrtInputFiles.txt ' +
+                              outdir + '/check.vrt'])
 
-        parameters['HOST'] = ''
-        self.assertEqual(obj.getConnectionString(parameters, context),
-                         "port=5432 active_schema=public")
+            cmd = alg.getConsoleCommands({'INPUT': [source],
+                                          'SRC_NODATA': '-9999 9999',
+                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
+            t = cmd[1]
+            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
+            self.assertEqual(cmd,
+                             ['gdalbuildvrt',
+                              '-resolution average -separate -r nearest -srcnodata "-9999 9999" ' +
+                              '-input_file_list buildvrtInputFiles.txt ' +
+                              outdir + '/check.vrt'])
 
-        parameters['PORT'] = '5555'
-        self.assertEqual(obj.getConnectionString(parameters, context),
-                         "port=5555 active_schema=public")
-
-        parameters['PORT'] = ''
-        self.assertEqual(obj.getConnectionString(parameters, context),
-                         "active_schema=public")
-
-        parameters['USER'] = 'usr'
-        self.assertEqual(obj.getConnectionString(parameters, context),
-                         "active_schema=public user=usr")
-
-        parameters['PASSWORD'] = 'pwd'
-        self.assertEqual(obj.getConnectionString(parameters, context),
-                         "password=pwd active_schema=public user=usr")
-
+            cmd = alg.getConsoleCommands({'INPUT': [source],
+                                          'SRC_NODATA': '',
+                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
+            t = cmd[1]
+            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
+            self.assertEqual(cmd,
+                             ['gdalbuildvrt',
+                              '-resolution average -separate -r nearest ' +
+                              '-input_file_list buildvrtInputFiles.txt ' +
+                              outdir + '/check.vrt'])
+            # additional parameters
+            cmd = alg.getConsoleCommands({'INPUT': [source],
+                                          'EXTRA': '-overwrite -optim RASTER -vrtnodata -9999',
+                                          'OUTPUT': outdir + '/check.vrt'}, context, feedback)
+            t = cmd[1]
+            cmd[1] = t[:t.find('-input_file_list') + 17] + t[t.find('buildvrtInputFiles.txt'):]
+            self.assertEqual(cmd,
+                             ['gdalbuildvrt',
+                              '-resolution average -separate -r nearest -overwrite -optim RASTER -vrtnodata -9999 ' +
+                              '-input_file_list buildvrtInputFiles.txt ' +
+                              outdir + '/check.vrt'])
 
 if __name__ == '__main__':
     nose2.main()
